+++ src/main.ts
@@ -1,4 +1,4 @@
-import BarGlob3d from './lib/barGlobe';
-import Glob3d from './lib/globe';
+import BarGlob3d from './lib/BarGlob3d';
+import Glob3d from './lib/Glob3d';
 
 export { BarGlob3d, Glob3d };
